Trigger discovery once installation is finished for such environments
